fix: uptake toast component
