+++ shared-helpers/index.ts
@@ -24,3 +24,4 @@ export * from "./src/views/layout/ExygyFooter"
 export * from "./src/views/sign-in/FormSignIn"
 export * from "./src/views/sign-in/FormSignInErrorBox"
 export * from "./src/views/sign-in/ResendConfirmationModal"
+export * from "./src/utilities/MessageContext"
