+++ sites/partners/cypress/e2e/default/07-jurisdictional-admin-user-management.spec.ts
@@ -50,7 +50,7 @@ describe("Jurisdictional Admin User Mangement Tests", () => {
       )
     })
     cy.getByTestId("invite-user").click()
-    cy.getByTestId("alert-box").contains("Invite sent").should("have.text", "Invite sent")
+    cy.getByTestId("toast-alert").contains("Invite sent").should("have.text", "Invite sent")
   })
 
   it("as jurisdictional admin user, should be able to create new partner", () => {
@@ -85,6 +85,6 @@ describe("Jurisdictional Admin User Mangement Tests", () => {
     cy.getByTestId("listings_Alameda").first().click()
     cy.getByTestId("listings_Alameda").last().click()
     cy.getByTestId("invite-user").click()
-    cy.getByTestId("alert-box").contains("Invite sent").should("have.text", "Invite sent")
+    cy.getByTestId("toast-alert").contains("Invite sent").should("have.text", "Invite sent")
   })
 })
