+++ sites/partners/src/lib/hooks.ts
@@ -3,7 +3,7 @@ import useSWR from "swr"
 import qs from "qs"
 import dayjs from "dayjs"
 import JSZip from "jszip"
-import { AuthContext } from "@bloom-housing/shared-helpers"
+import { AuthContext, MessageContext, alertTypes } from "@bloom-housing/shared-helpers"
 import {
   ApplicationSection,
   EnumApplicationsApiExtraModelOrder,
@@ -13,7 +13,7 @@ import {
   EnumUserFilterParamsComparison,
   UserRolesOnly,
 } from "@bloom-housing/backend-core/types"
-import { setSiteAlertMessage, t } from "@bloom-housing/ui-components"
+import { t } from "@bloom-housing/ui-components"
 export interface PaginationProps {
   page?: number
   limit: number | "all"
@@ -120,6 +120,7 @@ export function useListingsData({
 
 export const useListingZip = () => {
   const { listingsService } = useContext(AuthContext)
+  const { setToast } = useContext(MessageContext)
 
   const [zipExportLoading, setZipExportLoading] = useState(false)
   const [zipExportError, setZipExportError] = useState(false)
@@ -145,12 +146,12 @@ export const useListingZip = () => {
         fileLink.click()
       })
       setZipCompleted(true)
-      setSiteAlertMessage(t("t.exportSuccess"), "success")
+      setToast(t("t.exportSuccess"), { variant: "success", hideTimeout: 3000 })
     } catch (err) {
       setZipExportError(true)
     }
     setZipExportLoading(false)
-  }, [listingsService])
+  }, [listingsService, setToast])
 
   return {
     onExport,
@@ -548,6 +549,7 @@ const useCsvExport = (endpoint: () => Promise<string>, fileName: string) => {
   const [csvExportLoading, setCsvExportLoading] = useState(false)
   const [csvExportError, setCsvExportError] = useState(false)
   const [csvExportSuccess, setCsvExportSuccess] = useState(false)
+  const { setToast } = useContext(MessageContext)
 
   const onExport = useCallback(async () => {
     setCsvExportError(false)
@@ -562,14 +564,14 @@ const useCsvExport = (endpoint: () => Promise<string>, fileName: string) => {
       fileLink.href = URL.createObjectURL(blob)
       fileLink.click()
       setCsvExportSuccess(true)
-      setSiteAlertMessage(t("t.exportSuccess"), "success")
+      setToast(t("t.exportSuccess"), { variant: alertTypes.success, hideTimeout: 3000 })
     } catch (err) {
       console.log(err)
       setCsvExportError(true)
     }
 
     setCsvExportLoading(false)
-  }, [endpoint, fileName])
+  }, [endpoint, fileName, setToast])
 
   return {
     onExport,
