+++ sites/partners/styles/overrides.scss
@@ -13,4 +13,8 @@
     justify-content: center;
     text-align: center;
   }
+
+  #seeds-toast-stack {
+    --toast-stack-inset-top: 8.5rem;
+  }
 }
