+++ sites/public/src/pages/_app.tsx
@@ -9,6 +9,7 @@ import {
   LoggedInUserIdleTimeout,
   ConfigProvider,
   AuthProvider,
+  MessageProvider,
 } from "@bloom-housing/shared-helpers"
 import { headScript, bodyTopTag, pageChangeHandler } from "../lib/customScripts"
 import { AppSubmissionContext } from "../lib/applications/AppSubmissionContext"
@@ -92,7 +93,9 @@ function BloomApp({ Component, router, pageProps }: AppProps) {
         <ConfigProvider apiUrl={process.env.backendApiBase}>
           <AuthProvider>
             <LoggedInUserIdleTimeout onTimeout={() => conductor.reset()} />
-            <Component {...pageProps} />
+            <MessageProvider>
+              <Component {...pageProps} />
+            </MessageProvider>
           </AuthProvider>
         </ConfigProvider>
       </AppSubmissionContext.Provider>
