+++ sites/public/src/pages/reset-password.tsx
@@ -11,7 +11,6 @@ import {
   t,
   AlertBox,
   SiteAlert,
-  setSiteAlertMessage,
 } from "@bloom-housing/ui-components"
 import { PageView, pushGtmEvent, AuthContext } from "@bloom-housing/shared-helpers"
 import { UserStatus } from "../lib/constants"
@@ -44,8 +43,10 @@ const ResetPassword = () => {
 
     try {
       const user = await resetPassword(token.toString(), password, passwordConfirmation)
-      setSiteAlertMessage(t(`authentication.signIn.success`, { name: user.firstName }), "success")
-      await router.push("/account/applications")
+      await router.push({
+        pathname: "/account/dashboard",
+        query: { alert: t(`authentication.signIn.success`, { name: user.firstName }) },
+      })
     } catch (err) {
       const { status, data } = err.response || {}
       if (status === 400) {
